Server route to make discovery api call for SubjectHeading bibs
